+++ phoenix-scala/phoenix/app/phoenix/responses/cord/base/CordResponsePayments.scala
@@ -9,6 +9,7 @@ import phoenix.responses.AddressResponse
 import phoenix.responses.cord.base.CordResponseStoreCreditPayment.CordResponseApplePayPayment
 import phoenix.utils.aliases._
 import slick.jdbc.PostgresProfile.api._
+import core.utils.Money._
 
 sealed trait CordResponsePayments {
   def `type`: Type
@@ -60,9 +61,9 @@ object CordResponseCreditCardPayment {
 }
 
 case class CordResponseGiftCardPayment(code: String,
-                                       amount: Int,
-                                       currentBalance: Int,
-                                       availableBalance: Int,
+                                       amount: Long,
+                                       currentBalance: Long,
+                                       availableBalance: Long,
                                        createdAt: Instant,
                                        `type`: Type = GiftCard)
     extends CordResponsePayments
@@ -84,9 +85,9 @@ object CordResponseGiftCardPayment {
 }
 
 case class CordResponseStoreCreditPayment(id: Int,
-                                          amount: Int,
-                                          currentBalance: Int,
-                                          availableBalance: Int,
+                                          amount: Long,
+                                          currentBalance: Long,
+                                          availableBalance: Long,
                                           createdAt: Instant,
                                           `type`: Type = StoreCredit)
     extends CordResponsePayments
